feat: moved auth buttons to the Quiz fragment
