+++ app/src/main/res/layout/fragment_quiz.xml
@@ -7,16 +7,33 @@
     android:gravity="center"
     android:padding="20dp">
 
-    <Button
-        android:id="@+id/settings"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/settings"
-        android:layout_gravity="end"
-        android:textSize="11sp"
-        android:backgroundTint="@color/teal_700"
-        android:textStyle="bold"
-        />
+    <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="end">
+
+        <Button
+                android:id="@+id/auth_button"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="5dp"
+                android:textSize="11sp"
+                android:backgroundTint="@color/teal_700"
+                android:textStyle="bold"
+                />
+
+        <Button
+                android:id="@+id/settings"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="5dp"
+                android:text="@string/settings"
+                android:textSize="11sp"
+                android:backgroundTint="@color/teal_700"
+                android:textStyle="bold"
+                />
+
+    </LinearLayout>
 
     <LinearLayout
         android:layout_width="wrap_content"
